% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotQuantileDiff.R
\name{plotQuantileDiff}
\alias{plotQuantileDiff}
\title{Visualize the Difference between Quantiles of Two Classes}
\usage{
plotQuantileDiff(X, Y, theta = seq(0.02, 0.98, 0.02), method = c("2D",
  "3D"), plot = TRUE)
}
\arguments{
\item{X}{A \code{n*p} matrix containg \code{n} observations of \code{p} variables for training.}

\item{Y}{A vector of length \code{n} containing the class labels 1 or 2.}

\item{theta}{A vector of candidate probabilities of quantiles.}

\item{method}{Whether the differences are shown in 2-dimension or in 3-dimension.}

\item{plot}{whether there is a plot or not.}
}
\value{
A \code{p*T} matrix of the abosulte standarized quantile differences between two classes,
where \code{T} is the number of candidate quantiles.
}
\description{
Visualize the abosulte standarized quantile differences between two classes.
}
\examples{
# Read data
data(wdbc)
X <- as.matrix(wdbc[,-1])
Y <- wdbc[,1]

# Visualize in a 2D heatmap
plotQuantileDiff(X,Y)

# Visualize in 3D
plotQuantileDiff(X,Y,method="3D")

}
\references{
Lai
}
\author{
Yuanhao Lai
}
