% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skewFlip.R
\name{skewFlip}
\alias{skewFlip}
\title{Flip Signs of Variables based on Skewness Measures}
\usage{
skewFlip(X, cl = NULL, signSkew = NULL, skew.correct = c("Galton",
  "Kelley", "skewness", "none"))
}
\arguments{
\item{X}{A \code{n*p} matrix containg \code{n} observations of \code{p} variables.}

\item{cl}{A vector of length \code{n} containing the class labels 1,2,....}

\item{signSkew}{sign of each variable to have a non-negative skewness. Default is NULL and
it will be determined by the skewness measure on \code{X}.
If \code{signSkew} is specified, then \code{cl} and \code{skew.correct} will be ignored.}

\item{skew.correct}{Skewness measures applied to correct the skewness direction of the variables.
The possibile choices are: Galton's skewness (default),
Kelley's skewness, the conventional skewness index based on the third standardized moment and
no correction..}
}
\value{
A list of the following two components:

\item{X}{A \code{n*p} matrix with the sign of variables flipped to have positive skewnesses.}
\item{signSkew}{A vector of length \code{p} containing flip sign of each variable.}
}
\description{
Flip the signs of variables based on the specified skewness measures to ensure that
all varibales have non-negative skewnesses.
}
\examples{
# Divide data into training set and test set randomly
data(wdbc)
set.seed(193)
trainIndex <- sample(c(rep(TRUE,2),rep(FALSE,1)),nrow(wdbc),replace=TRUE)
train <- as.matrix(wdbc[trainIndex,-1])
cl.train <- wdbc[trainIndex,1]
test <- as.matrix(wdbc[!trainIndex,-1])
cl.test <- wdbc[!trainIndex,1]

# Flip signs of variables to ensure same sign of skewness in the traing set
trainFlip <- skewFlip(train,cl.train,skew.correct="Galton")
signSkew <- trainFlip$signSkew
trainFlip <- trainFlip$X

# Apply the sign above to flip the test set
testFlip <- skewFlip(test,signSkew=signSkew)$X


}
\references{
C. Hennig and  C. Viroli. Quantile-based classifiers.Biometrika, 103(2):435–446, 2016.
}
\author{
Yuanhao Lai
}
