% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skewFlip.R
\name{skewFun}
\alias{skewFun}
\title{Skewness Measure}
\usage{
skewFun(x, skew.type = c("Galton", "Kelley", "skewness"))
}
\arguments{
\item{x}{A vector.}

\item{skew.type}{Skewness measures applied to correct the skewness direction of the variables.
The possibile choices are: Galton's skewness (default),
Kelley's skewness and the conventional skewness index based on the third standardized moment.}
}
\value{
A value of the specified skewness measure.
}
\description{
Provide the computation of the Galton’s skewness measure, the Kelley skewness measure,
or the convention skewness measure.
}
\examples{
x <- rexp(100)
skewFun(x,skew.type="Galton")


}
\references{
C. Hennig and  C. Viroli. Quantile-based classifiers.Biometrika, 103(2):435–446, 2016.
}
\author{
Yuanhao Lai
}
\keyword{internal}
